small refactor and error handling improvements

+++ pages/api/gallery/delete.ts
@@ -4,14 +4,19 @@ import { unstable_getServerSession } from 'next-auth'
 import { authOptions } from '../auth/[...nextauth]'
 import { prisma } from 'lib/prisma'
 
+type Data = {
+  message: string
+  error?: boolean
+}
+
 export default async function handler(
   req: NextApiRequest,
-  res: NextApiResponse
+  res: NextApiResponse<Data>
 ) {
   const session = await unstable_getServerSession(req, res, authOptions)
 
   if (session && session.user.role === 'ADMIN') {
-    if (req.method === 'DELETE') {
+    if (req.method === 'DELETE' && req.query.itemId) {
       try {
         const item = await prisma.item.delete({
           where: { itemId: req.query.itemId as string },
@@ -23,12 +28,16 @@ export default async function handler(
           .status(200)
           .json({ message: `itemId ${item.itemId} has been deleted!` })
       } catch (error) {
-        return res.status(422).json(error)
+        return res.status(422).json({
+          message: `Something went wrong. itemId ${req.query.itemId} was not deleted.`,
+          error: true,
+        })
       }
     }
   }
 
-  res.status(401).send({
-    error: 'You must be an admin to view the protected content.',
+  res.status(401).json({
+    message: 'You must be an admin to view the protected content.',
+    error: true,
   })
 }
